test(text): split GSUB context regressions
